# Conflicts:
#	frontend/src/business/components/track/case/components/TestCaseEditOtherInfo.vue
#	frontend/src/business/components/track/plan/view/comonents/report/detail/TestPlanReportButtons.vue

+++ backend/src/main/java/io/metersphere/api/service/ApiScenarioReportService.java
@@ -32,7 +32,6 @@ import io.metersphere.utils.LoggerUtil;
 import org.apache.commons.beanutils.BeanMap;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.ibatis.session.SqlSessionFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -78,7 +77,7 @@ public class ApiScenarioReportService {
     @Resource
     private ApiScenarioReportStructureMapper apiScenarioReportStructureMapper;
     @Resource
-    private SqlSessionFactory sqlSessionFactory;
+    private ApiDefinitionExecResultMapper definitionExecResultMapper;
 
     public void saveResult(List<RequestResult> requestResults, ResultDTO dto) {
         // 报告详情内容
@@ -118,7 +117,7 @@ public class ApiScenarioReportService {
         APIScenarioReportResult reportResult = extApiScenarioReportMapper.get(reportId);
         if (reportResult != null) {
             if (reportResult.getReportVersion() != null && reportResult.getReportVersion() > 1) {
-                reportResult.setContent(JSON.toJSONString(apiScenarioReportStructureService.getReport(reportId)));
+                reportResult.setContent(JSON.toJSONString(apiScenarioReportStructureService.assembleReport(reportId)));
             } else {
                 ApiScenarioReportDetail detail = apiScenarioReportDetailMapper.selectByPrimaryKey(reportId);
                 if (detail != null && reportResult != null) {
@@ -319,17 +318,25 @@ public class ApiScenarioReportService {
         return report;
     }
 
-    public void margeReport(String reportId) {
+    public void margeReport(String reportId, String runMode) {
         ApiScenarioReport report = apiScenarioReportMapper.selectByPrimaryKey(reportId);
         if (report != null) {
             // 更新场景状态
-            ApiScenarioReportResultExample example = new ApiScenarioReportResultExample();
-            example.createCriteria().andReportIdEqualTo(reportId).andStatusEqualTo(ScenarioStatus.Error.name());
-            long size = apiScenarioReportResultMapper.countByExample(example);
-            report.setStatus(size > 0 ? ScenarioStatus.Error.name() : ScenarioStatus.Success.name());
-            report.setEndTime(System.currentTimeMillis());
+            if (StringUtils.equalsIgnoreCase(runMode, ApiRunMode.DEFINITION.name())) {
+                ApiDefinitionExecResultExample execResultExample = new ApiDefinitionExecResultExample();
+                execResultExample.createCriteria().andIntegratedReportIdEqualTo(reportId).andStatusEqualTo("Error");
+                long size = definitionExecResultMapper.countByExample(execResultExample);
+                report.setStatus(size > 0 ? ScenarioStatus.Error.name() : ScenarioStatus.Success.name());
+            } else {
+                ApiScenarioReportResultExample example = new ApiScenarioReportResultExample();
+                example.createCriteria().andReportIdEqualTo(reportId).andStatusEqualTo(ScenarioStatus.Error.name());
+                long size = apiScenarioReportResultMapper.countByExample(example);
+                report.setStatus(size > 0 ? ScenarioStatus.Error.name() : ScenarioStatus.Success.name());
+            }
             // 更新控制台信息
             apiScenarioReportStructureService.update(reportId, FixedCapacityUtils.getJmeterLogger(reportId));
+
+            report.setEndTime(System.currentTimeMillis());
             // 更新报告
             apiScenarioReportMapper.updateByPrimaryKey(report);
         }
@@ -471,6 +478,14 @@ public class ApiScenarioReportService {
         structureExample.createCriteria().andReportIdEqualTo(request.getId());
         apiScenarioReportStructureMapper.deleteByExample(structureExample);
 
+        ApiDefinitionExecResultExample definitionExecResultExample = new ApiDefinitionExecResultExample();
+        definitionExecResultExample.createCriteria().andIdEqualTo(request.getId());
+        definitionExecResultMapper.deleteByExample(definitionExecResultExample);
+
+        ApiDefinitionExecResultExample execResultExample = new ApiDefinitionExecResultExample();
+        execResultExample.createCriteria().andIntegratedReportIdEqualTo(request.getId());
+        definitionExecResultMapper.deleteByExample(execResultExample);
+
         // 补充逻辑，如果是集成报告则把零时报告全部删除
         ApiScenarioReport report = apiScenarioReportMapper.selectByPrimaryKey(request.getId());
         if (report != null && StringUtils.isNotEmpty(report.getScenarioId())) {
@@ -487,15 +502,49 @@ public class ApiScenarioReportService {
     public void delete(String id) {
         apiScenarioReportDetailMapper.deleteByPrimaryKey(id);
         apiScenarioReportMapper.deleteByPrimaryKey(id);
+        ApiScenarioReportResultExample example = new ApiScenarioReportResultExample();
+        example.createCriteria().andReportIdEqualTo(id);
+        apiScenarioReportResultMapper.deleteByExample(example);
+
+        ApiScenarioReportStructureExample structureExample = new ApiScenarioReportStructureExample();
+        structureExample.createCriteria().andReportIdEqualTo(id);
+        apiScenarioReportStructureMapper.deleteByExample(structureExample);
+
+        ApiDefinitionExecResultExample definitionExecResultExample = new ApiDefinitionExecResultExample();
+        definitionExecResultExample.createCriteria().andIdEqualTo(id);
+        definitionExecResultMapper.deleteByExample(definitionExecResultExample);
+
+        ApiDefinitionExecResultExample execResultExample = new ApiDefinitionExecResultExample();
+        execResultExample.createCriteria().andIntegratedReportIdEqualTo(id);
+        definitionExecResultMapper.deleteByExample(execResultExample);
+
     }
 
     public void deleteByIds(List<String> ids) {
-        ApiScenarioReportExample example = new ApiScenarioReportExample();
-        example.createCriteria().andIdIn(ids);
-        ApiScenarioReportDetailExample detailExample = new ApiScenarioReportDetailExample();
-        detailExample.createCriteria().andReportIdIn(ids);
-        apiScenarioReportDetailMapper.deleteByExample(detailExample);
-        apiScenarioReportMapper.deleteByExample(example);
+        if (CollectionUtils.isNotEmpty(ids)) {
+            ApiScenarioReportExample example = new ApiScenarioReportExample();
+            example.createCriteria().andIdIn(ids);
+            ApiScenarioReportDetailExample detailExample = new ApiScenarioReportDetailExample();
+            detailExample.createCriteria().andReportIdIn(ids);
+            apiScenarioReportDetailMapper.deleteByExample(detailExample);
+            apiScenarioReportMapper.deleteByExample(example);
+
+            ApiScenarioReportResultExample reportResultExample = new ApiScenarioReportResultExample();
+            reportResultExample.createCriteria().andReportIdIn(ids);
+            apiScenarioReportResultMapper.deleteByExample(reportResultExample);
+
+            ApiScenarioReportStructureExample structureExample = new ApiScenarioReportStructureExample();
+            structureExample.createCriteria().andReportIdIn(ids);
+            apiScenarioReportStructureMapper.deleteByExample(structureExample);
+
+            ApiDefinitionExecResultExample definitionExecResultExample = new ApiDefinitionExecResultExample();
+            definitionExecResultExample.createCriteria().andIdIn(ids);
+            definitionExecResultMapper.deleteByExample(definitionExecResultExample);
+
+            ApiDefinitionExecResultExample execResultExample = new ApiDefinitionExecResultExample();
+            execResultExample.createCriteria().andIntegratedReportIdIn(ids);
+            definitionExecResultMapper.deleteByExample(execResultExample);
+        }
     }
 
     public void deleteAPIReportBatch(APIReportBatchRequest reportRequest) {
@@ -549,6 +598,14 @@ public class ApiScenarioReportService {
             structureExample.createCriteria().andReportIdIn(handleIdList);
             apiScenarioReportStructureMapper.deleteByExample(structureExample);
 
+            ApiDefinitionExecResultExample definitionExecResultExample = new ApiDefinitionExecResultExample();
+            definitionExecResultExample.createCriteria().andIdIn(handleIdList);
+            definitionExecResultMapper.deleteByExample(definitionExecResultExample);
+
+            ApiDefinitionExecResultExample execResultExample = new ApiDefinitionExecResultExample();
+            execResultExample.createCriteria().andIntegratedReportIdIn(handleIdList);
+            definitionExecResultMapper.deleteByExample(execResultExample);
+
             //转存剩余的数据
             ids = otherIdList;
         }
@@ -570,6 +627,13 @@ public class ApiScenarioReportService {
             structureExample.createCriteria().andReportIdIn(ids);
             apiScenarioReportStructureMapper.deleteByExample(structureExample);
 
+            ApiDefinitionExecResultExample definitionExecResultExample = new ApiDefinitionExecResultExample();
+            definitionExecResultExample.createCriteria().andIdIn(ids);
+            definitionExecResultMapper.deleteByExample(definitionExecResultExample);
+
+            ApiDefinitionExecResultExample execResultExample = new ApiDefinitionExecResultExample();
+            execResultExample.createCriteria().andIntegratedReportIdIn(ids);
+            definitionExecResultMapper.deleteByExample(execResultExample);
         }
     }
 
@@ -695,6 +759,7 @@ public class ApiScenarioReportService {
         report.setProjectId(projectId);
         report.setScenarioName(scenarioName);
         report.setScenarioId(scenarioId);
+        report.setReportType(ReportTypeConstants.SCENARIO_INDEPENDENT.name());
         return report;
     }
 
@@ -734,7 +799,10 @@ public class ApiScenarioReportService {
         List<ApiScenarioReport> apiScenarioReports = apiScenarioReportMapper.selectByExample(example);
         List<String> ids = apiScenarioReports.stream().map(ApiScenarioReport::getId).collect(Collectors.toList());
         if (CollectionUtils.isNotEmpty(ids)) {
-            deleteByIds(ids);
+            APIReportBatchRequest request = new APIReportBatchRequest();
+            request.setIds(ids);
+            request.setSelectAllDate(false);
+            deleteAPIReportBatch(request);
         }
     }
 
@@ -753,11 +821,19 @@ public class ApiScenarioReportService {
         }
     }
 
-    public void reName(QueryAPIReportRequest reportRequest) {
-        ApiScenarioReport apiTestReport = apiScenarioReportMapper.selectByPrimaryKey(reportRequest.getId());
-        if (apiTestReport != null) {
-            apiTestReport.setName(reportRequest.getName());
-            apiScenarioReportMapper.updateByPrimaryKey(apiTestReport);
+    public void reName(ApiScenarioReport reportRequest) {
+        if (StringUtils.equalsIgnoreCase(reportRequest.getReportType(), ReportTypeConstants.API_INDEPENDENT.name())) {
+            ApiDefinitionExecResult result = definitionExecResultMapper.selectByPrimaryKey(reportRequest.getId());
+            if (result != null) {
+                result.setName(reportRequest.getName());
+                definitionExecResultMapper.updateByPrimaryKeySelective(result);
+            }
+        } else {
+            ApiScenarioReport apiTestReport = apiScenarioReportMapper.selectByPrimaryKey(reportRequest.getId());
+            if (apiTestReport != null) {
+                apiTestReport.setName(reportRequest.getName());
+                apiScenarioReportMapper.updateByPrimaryKey(apiTestReport);
+            }
         }
     }
 }
